Update Bestiary (Lightning Dragons & Copper)

+++ src/main/java/com/github/alexthe666/iceandfire/client/gui/bestiary/IndexPageButton.java
@@ -25,7 +25,7 @@ public class IndexPageButton extends GuiButton {
 			mc.renderEngine.bindTexture(new ResourceLocation("iceandfire:textures/gui/bestiary/widgets.png"));
 			this.hovered = mouseX >= this.x && mouseY >= this.y && mouseX < this.x + this.width && mouseY < this.y + this.height;
 			this.drawTexturedModalRect(this.x, this.y, 0, this.hovered ? 32 : 0, this.width, this.height);
-			fontrenderer.drawString(this.displayString, (float) (this.x + this.width / 2 - fontrenderer.getStringWidth(this.displayString) / 2), (float) this.y + (this.height - 8) / 2, this.hovered ? 0XFAE67D : 0X303030, false);
+			fontrenderer.drawString(this.displayString, (float) (this.x + this.width / 2 - fontrenderer.getStringWidth(this.displayString) / 2), (float) this.y + (float) (this.height - 8) / 2, this.hovered ? 0XFAE67D : 0X303030, false);
 		}
 	}
 }
